Introduce ObjectDatabase.MergeCommits
Provide an API to merge two commits and return the resultant Index.
This Index may be analyzed looking for conflicts or otherwise to
determine the success of the merge.  This Index must be disposed when
the caller has finished with it.  Thus, Index now implements
IDisposable and Indexes that are not the repository index (ie, generated
by merge) are expected to be disposed.  The repository's index is
not expected to be disposed and is protected from disposal.